opt/requirements.txt
azure-eventgrid==4.1.1
flask==1.1.2
python-dotenv==0.17.0
paho-mqtt==1.5.1
